+++ tests/it/main.rs
@@ -1,6 +1,7 @@
 mod basic;
 mod errors;
 mod generic;
+mod matching_eq;
 mod mixed;
 mod mock_order;
 mod prefix;