+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `api=[..]` attribute option for opting out of generating an encapsulating module.
 - Support for self references in `Option` and `Result` return values (e.g. `-> Option<&str>`)
 - Better mismatch debugging with improved `matching!` API under the hood.
+- Support for `Eq`-based matching in the `matching!` macro.
 ### Fixed
 - Avoid spitting out `#[allow(non_camel_case_types)]` when the user supplied the MockFn type names.
 ### Removed